code_samples/deployment/02_production_api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
prometheus-client==0.19.0
python-dotenv==1.0.0
structlog==23.2.0